Attempt to refactor tta_need components into own file tree
