+++ app/javascript/monitoring_report/reducers/app.js
@@ -1,13 +1,10 @@
 import {
-  SET_APP_CONFIG,
   TOGGLE_REQUEST_FORM,
   SHOW_SUCCESS_MESSAGE
-} from '../actions'
+} from 'tta_need/actions'
 
 const app = (state = {}, action) => {
   switch(action.type) {
-    case SET_APP_CONFIG:
-      return {...state, ...action.config}
     case TOGGLE_REQUEST_FORM:
       return {...state, formOpen: action.value, showSuccess: false}
     case SHOW_SUCCESS_MESSAGE:
