+++ app/javascript/monitoring_report/reducers/index.js
@@ -2,8 +2,7 @@ import { combineReducers } from 'redux'
 import { reducer as bees } from 'redux-bees'
 import app from './app'
 import report from './report'
-import ttaNeed from './tta_need'
-import tasks from './tasks'
+import { ttaNeed, tasks } from 'tta_need/reducers'
 
 export default combineReducers({
   app,
