+++ app/javascript/packs/monitoring_report.jsx
@@ -3,15 +3,12 @@ import { render } from 'react-dom'
 import { Provider } from 'react-redux'
 import MonitoringReportApp from 'monitoring_report/app'
 import store from 'monitoring_report/store'
-import { setAppConfig, setReportId } from 'monitoring_report/actions'
+import { setReportId } from 'monitoring_report/actions'
 import 'core-js'
 import 'regenerator-runtime/runtime'
 import 'whatwg-fetch'
 
 const target = document.getElementById("monitoring-report")
-store.dispatch(setAppConfig({
-  endpoint: window.location.origin
-}))
 store.dispatch(setReportId(target.getAttribute("data-report-id")))
 
 render(
