+++ app/javascript/tta_need/actions/index.js
@@ -1,0 +1,2 @@
+export * from './tta_need'
+export * from './tasks'
