+++ app/javascript/tta_need/api.js
@@ -2,7 +2,9 @@ import { buildApi, get, post } from 'redux-bees'
 import camelcaseKeys from 'camelcase-keys'
 
 const apiEndpoints = {
-  getMonitoringReport: { method: get, path: "/monitoring_reports/:id" }
+  getTopics: { method: get, path: "/topics" },
+  createNeed: { method: post, path: "/grantees/:granteeId/tta_needs" },
+  createTask: { method: post, path: "/tta_needs/:ttaNeedId/tasks" }
 }
 
 const config = {
