+++ app/javascript/tta_need/reducers/index.js
@@ -1,0 +1,2 @@
+export { default as ttaNeed } from './tta_need'
+export { default as tasks } from './tasks'
