+++ app/javascript/tta_need/reducers/tta_need.js
@@ -1,14 +1,12 @@
-import {
-  UPDATE_NEED_FIELDS,
-  SET_REPORT_ID
-} from '../actions'
+import { UPDATE_NEED_FIELDS } from '../actions'
+import { SET_REPORT_ID } from 'monitoring_report/actions'
 
 const ttaNeed = (state = {}, action) => {
   switch(action.type) {
     case UPDATE_NEED_FIELDS:
       return {...state, ...action.fields}
     case SET_REPORT_ID:
-      return {...state, contextLinkId: action.id}
+      return {...state, contextLinkType: "monitoring-reports", contextLinkId: action.id}
     default:
       return state
   }
